# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ADDABAAZ API
  version: 2.0.0
  description: |
    Backend for the ADDABAAZ web and mobile apps. Base path `/api/v1`.
    Auth is a bearer JWT (`Authorization: Bearer <token>`) returned by signup/login.
    Errors are `{ "error": { "code": "...", "message": "..." } }`. Write endpoints return `204` unless noted.
servers:
  - url: http://localhost:3000/api/v1
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
    admin: { type: http, scheme: bearer, description: 'An administrator account’s session token (users.is_admin; valid 12h after sign-in), or the optional ADMIN_TOKEN environment variable (24+ characters) for scripts' }
...
security: [ { bearer: [] } ]
paths:
...
  /invoices/{id}/pdf:
    get:
      summary: 'Download a GST invoice or credit note (owner only)'
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { '200': { description: PDF, content: { application/pdf: { schema: { type: string, format: binary } } } }, '404': { description: not found } }
...
  /admin/payments:
    get: { summary: 'Admin (Bearer ADMIN_TOKEN): recent payments, ?email=&limit=', security: [ { admin: [] } ], responses: { '200': { description: '{ payments }' }, '401': { description: unauthorized }, '403': { description: not an administrator } } }
  /admin/payments/{id}/refund:
    post:
      summary: 'Admin: refund a Razorpay payment (fully or partly). Issues a credit note once processed; a full refund ends the access.'
      security: [ { admin: [] } ]
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody: { content: { application/json: { schema: { type: object, properties: { amountPaise: { type: integer, minimum: 100 }, reason: { type: string }, revokeAccess: { type: boolean } } } } } }
      responses: { '201': { description: '{ refund, creditNote, accessRevoked }' }, '409': { description: not_refundable | refund_exceeds_payment }, '502': { description: provider_error } }
  /admin/coupons:
    get: { summary: 'Admin: list coupons with redemption counts', security: [ { admin: [] } ], responses: { '200': { description: '{ coupons }' } } }
    post:
      summary: 'Admin: create a coupon'
      security: [ { admin: [] } ]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [code, kind, value], properties: { code: { type: string }, kind: { type: string, enum: [percent, flat] }, value: { type: integer, description: 'percent 1-100, or paise' }, description: { type: string }, planIds: { type: array, items: { type: string } }, maxRedemptions: { type: integer }, perUserLimit: { type: integer, default: 1 }, startsAt: { type: string, format: date-time }, expiresAt: { type: string, format: date-time } } } } } }
      responses: { '201': { description: '{ coupon }' }, '400': { description: validation }, '409': { description: exists } }
  /admin/coupons/{code}:
    patch: { summary: 'Admin: change active / dates / limits / description', security: [ { admin: [] } ], parameters: [ { name: code, in: path, required: true, schema: { type: string } } ], responses: { '200': { description: '{ coupon }' }, '404': { description: unknown } } }
  /admin/invoices.csv:
    get: { summary: 'Admin: sales register CSV (invoices +, credit notes -), ?from=YYYY-MM-DD&to=YYYY-MM-DD in IST', security: [ { admin: [] } ], responses: { '200': { description: text/csv } } }
  /admin/invoices/{id}/pdf:
    get: { summary: 'Admin: any invoice as PDF', security: [ { admin: [] } ], parameters: [ { name: id, in: path, required: true, schema: { type: string } } ], responses: { '200': { description: PDF } } }
  /admin/session:
    get: { summary: 'Admin: who am I (also proves the session is an admin one)', security: [ { admin: [] } ], responses: { '200': { description: ok }, '401': { description: unauthorized / admin_session_expired }, '403': { description: not an administrator } } }
  /admin/stats:
    get: { summary: 'Admin: dashboard numbers (revenue, subscribers, sign-ups, recent activity)', security: [ { admin: [] } ], responses: { '200': { description: ok }, '401': { description: unauthorized / admin_session_expired }, '403': { description: not an administrator } } }
...
  /admin/users:
    get: { summary: 'Admin: users, ?q=&filter=all|paid|free|expiring|expired|admin|disabled&limit=&offset=', security: [ { admin: [] } ], responses: { '200': { description: ok }, '401': { description: unauthorized / admin_session_expired }, '403': { description: not an administrator } } }
  /admin/users/{id}:
    get: { summary: 'Admin: one user with profiles, subscription and payments', security: [ { admin: [] } ], responses: { '200': { description: ok }, '401': { description: unauthorized / admin_session_expired }, '403': { description: not an administrator }, '404': { description: unknown user } } }
    patch: { summary: 'Admin: change name / isAdmin / disabled (cannot lock yourself out or remove the last admin)', security: [ { admin: [] } ], responses: { '200': { description: ok }, '401': { description: unauthorized / admin_session_expired }, '403': { description: not an administrator }, '409': { description: cannot_lock_yourself_out | last_admin } } }
    delete: { summary: 'Admin: delete a user (payment and invoice records are kept)', security: [ { admin: [] } ], responses: { '200': { description: ok }, '401': { description: unauthorized / admin_session_expired }, '403': { description: not an administrator }, '409': { description: cannot_lock_yourself_out | last_admin } } }
  /admin/users/{id}/grant:
    post: { summary: 'Admin: give free access { days, planId?, note? } — no payment or invoice', security: [ { admin: [] } ], responses: { '200': { description: ok }, '401': { description: unauthorized / admin_session_expired }, '403': { description: not an administrator } } }
  /admin/users/{id}/revoke-plan:
    post: { summary: 'Admin: end the current plan now (does not refund)', security: [ { admin: [] } ], responses: { '200': { description: ok }, '401': { description: unauthorized / admin_session_expired }, '403': { description: not an administrator } } }
  /admin/messages:
    get: { summary: 'Admin: contact messages, ?status=open|handled|all', security: [ { admin: [] } ], responses: { '200': { description: ok }, '401': { description: unauthorized / admin_session_expired }, '403': { description: not an administrator } } }
  /admin/messages/{id}:
    patch: { summary: 'Admin: mark handled / reopen { handled }', security: [ { admin: [] } ], responses: { '200': { description: ok }, '401': { description: unauthorized / admin_session_expired }, '403': { description: not an administrator } } }
    delete: { summary: 'Admin: delete a message', security: [ { admin: [] } ], responses: { '200': { description: ok }, '401': { description: unauthorized / admin_session_expired }, '403': { description: not an administrator } } }
  /admin/audit:
    get: { summary: 'Admin: audit log, newest first, ?action=&before=<id>&limit=', security: [ { admin: [] } ], responses: { '200': { description: ok }, '401': { description: unauthorized / admin_session_expired }, '403': { description: not an administrator } } }
  /admin/catalog:
    get: { summary: 'Admin: the whole catalog including the studio profile', security: [ { admin: [] } ], responses: { '200': { description: ok }, '401': { description: unauthorized / admin_session_expired }, '403': { description: not an administrator } } }
  /admin/catalog/{type}:
    post: { summary: 'Admin: add an item; type = shows|videos|upcoming|gallery. Validated like `npm run validate:catalog`', security: [ { admin: [] } ], responses: { '200': { description: ok }, '401': { description: unauthorized / admin_session_expired }, '403': { description: not an administrator }, '400': { description: validation }, '409': { description: exists } } }
  /admin/catalog/{type}/order:
    put: { summary: 'Admin: set display order { ids: [...] }', security: [ { admin: [] } ], responses: { '200': { description: ok }, '401': { description: unauthorized / admin_session_expired }, '403': { description: not an administrator } } }
  /admin/catalog/{type}/{id}:
    put: { summary: 'Admin: replace an item', security: [ { admin: [] } ], responses: { '200': { description: ok }, '401': { description: unauthorized / admin_session_expired }, '403': { description: not an administrator }, '404': { description: unknown } } }
    delete: { summary: 'Admin: delete an item (shows: ?cascade=1 also deletes their videos)', security: [ { admin: [] } ], responses: { '200': { description: ok }, '401': { description: unauthorized / admin_session_expired }, '403': { description: not an administrator }, '409': { description: has_videos } } }
  /admin/studio:
    put: { summary: 'Admin: replace the studio profile (About / Services / Contact)', security: [ { admin: [] } ], responses: { '200': { description: ok }, '401': { description: unauthorized / admin_session_expired }, '403': { description: not an administrator }, '400': { description: validation } } }
  /admin/uploads/image:
    post: { summary: 'Admin: upload an image (raw body, png/jpeg/webp/gif, ≤ 4 MB) → { path: uploads/<hash>.<ext> }', security: [ { admin: [] } ], responses: { '200': { description: ok }, '401': { description: unauthorized / admin_session_expired }, '403': { description: not an administrator }, '413': { description: too large }, '400': { description: unsupported_image } } }
  /admin/uploads/video:
    post: { summary: 'Admin: presigned PUT URL for a premium video in R2 { filename, slug } → { key, uploadUrl, expiresInSeconds }', security: [ { admin: [] } ], responses: { '200': { description: ok }, '401': { description: unauthorized / admin_session_expired }, '403': { description: not an administrator }, '503': { description: storage_not_configured } } }